Reservation

+++ src/app/components/admin/gestiondestination/gestiondestination.component.ts
@@ -18,6 +18,13 @@ export class GestiondestinationComponent implements OnInit {
   destination: Dentination = new Dentination(' ', ' ', ' ', ' ', 0, 0, 0);
   hotel1: Hotel = new Hotel(' ', ' ', ' ', 0, 0, 0, '', true);
 
+  reloadCurrentRoute() {
+    const currentUrl = this.router.url;
+    this.router.navigateByUrl('/', { skipLocationChange: true }).then(() => {
+      this.router.navigate([currentUrl]);
+    });
+  }
+
   getcityhotels(city: String) {
     this.hotelsService.getHotels().subscribe(
       (response) => {
@@ -62,7 +69,7 @@ export class GestiondestinationComponent implements OnInit {
     this.hotelsService.getOneHotel(hotel).subscribe(
       (response) => {
         this.hotel1 = response;
-        this.hotel1.enpromo=this.hotel1.promotion!=0
+        this.hotel1.enpromo = this.hotel1.promotion != 0;
         return this.hotel1;
       },
       (error) => {
@@ -72,20 +79,28 @@ export class GestiondestinationComponent implements OnInit {
   }
 
   destinationDelete(name: string) {
-    this.DentinationService.deleteDentination(name).subscribe(
-      (response) => {
-        this.SuccessSnackBar('Destination Deleted successfully');
-      },
-      (error) => {
-        this.ErrorSnackBar('Failed Modification');
-      }
-    );
+    if (this.cityhotels.length != 0) {
+      this.ErrorSnackBar(
+        'This Destination have some hotels please remove them before romoving the destination'
+      );
+    } else {
+      this.DentinationService.deleteDentination(name).subscribe(
+        (response) => {
+          this.SuccessSnackBar('Destination Deleted successfully');
+          this.reloadCurrentRoute();
+        },
+        (error) => {
+          this.ErrorSnackBar('Failed Modification');
+        }
+      );
+    }
   }
 
   hotelDelete(name: string) {
     this.hotelsService.deleteHotel(name).subscribe(
       (response) => {
         this.SuccessSnackBar('Hotel Deleted successfully');
+        this.reloadCurrentRoute();
       },
       (error) => {
         this.ErrorSnackBar('Failed Modification');
